requirements.txt
python-dotenv
pandas
numpy
matplotlib
pymongo[srv]==3.11
pymongo
certifi
kaggle
# -e .